+++ packages/extras/src/lib/index.ts
@@ -75,11 +75,6 @@ export {
   type IntersectionEvent
 } from './interactivity'
 
-// transitions
-export { transitions } from './transitions/transitions-plugin'
-export type { ThrelteTransition } from './transitions/types'
-export { createTransition } from './transitions/createTransition'
-
 // layers
 export { layers, type ThrelteLayers, type ThrelteLayersContext } from './layers'