+++ packages/xr/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @threlte/xr
 
+## 1.0.0-next.6
+
+### Patch Changes
+
+- cc4ccf2: Update Three.js
+
+## 1.0.0-next.5
+
+### Patch Changes
+
+- fa8a61c: Bump Svelte compiler version
+
 ## 1.0.0-next.4
 
 ### Patch Changes